app/requirements.txt
fastapi==0.70.0
uvicorn
sqlalchemy==1.4.28
bcrypt==3.2.0
pydantic
typing
python-multipart
python-jose[cryptography]
passlib[bcrypt]
